Refactor: Add logo property to node types
This commit adds a new `logo` property to the `NodeTypeSchema` class in the `node_type_schemas.py` file. The `logo` property is used to specify the logo for a node type. If a logo is available for a node type, it will be displayed alongside the node title in the frontend.

The `logo` property is also added to the `BaseNode` class in the `base.py` file. The `logo` property is set to `None` by default and can be overridden in subclasses.

Additionally, the `SlackNode` class in the `slack.py` file is updated to include a `logo` property with the value "/images/slack.png". This logo will be displayed for Slack nodes in the frontend.

Lastly, the `BaseNode` component in the `BaseNode.tsx` file is modified to display the node logo if available. The logo is fetched from the `PUBLIC_URL` and displayed alongside the node title.

This refactoring improves the visual representation of node types in the frontend and allows for better customization and branding of node types.